-updated mod version -update game version -removed cloud 'generation'(like anuke puddle)

+++ tools/src/gas/tools/AnukeCompDownloader.java
@@ -11,23 +11,28 @@ import org.apache.commons.io.FileUtils;
 import java.net.URL;
 
 public class AnukeCompDownloader {
+    private static final String packageName = "gas", annotationsClassName = "GasAnnotations";
+    private static final JavaCodeConverter codeConverter = new JavaCodeConverter(false);
     private static String selectedClassName = "";
-    private static JavaCodeConverter codeConverter=new JavaCodeConverter();
 
     public static void main(String[] args) {
-        Log.info("[@]", Seq.with(args).toString(", "));
+        String mindustryVersion = Seq.with(args).find(s -> s.startsWith("v"));
+        if (mindustryVersion==null){
+            System.out.println("Please put mindustry version in args!!!");
+            System.exit(1);
+            return;
+        }
+        Log.info("Checking Anuke's comps for "+mindustryVersion);
         Fi folder = new Fi("debug");
         try {
             long nanos = System.nanoTime();
-            Fi dir = new Fi("core/src/gas/entities/compByAnuke");
+            Fi dir = new Fi("core/src/" + packageName + "/entities/compByAnuke");
 
-            String mindustryVersion = Seq.with(args).find(s -> s.startsWith("v"));
             if (dir.exists()) dir.delete();
-//            dir.mkdirs();
             folder.mkdirs();
             Fi compJava = folder.child("compJava");
             Fi finalComp = folder.child("finalComp");
-            boolean downloadNew = false, log = false;
+            boolean downloadNew = false;
 
             Fi version = new Fi("compDownloader").child("version.txt");
             Fi sourcesFi = new Fi("compDownloader").child("sources.zip");
@@ -35,15 +40,14 @@ public class AnukeCompDownloader {
                 downloadNew = true;
                 version.writeString(mindustryVersion);
             }
-//        URL sourceZipLink = ;https://codeload.github.com/Anuken/Mindustry/zip/refs/tags/v126.2
-
             if (downloadNew || !sourcesFi.exists()) {
-                Log.info("downloading new version");
+                Log.info("Downloading new comps version");
                 Time.mark();
                 FileUtils.copyURLToFile(new URL("https://codeload.github.com/Anuken/Mindustry/zip/refs/tags/" + mindustryVersion), sourcesFi.file(), 10000, 10000);
                 Log.info("Time to download: @ms", Time.elapsed());
+            } else {
+                Log.info("Game version and comps version are the same");
             }
-
             ZipFi sourceZip = new ZipFi(sourcesFi);
 
             Fi child = sourceZip.list()[0].child("core").child("src").child("mindustry").child("entities").child("comp");
@@ -51,20 +55,10 @@ public class AnukeCompDownloader {
                 fi.copyTo(compJava.child(fi.name()));
             }
             for (Fi fi : compJava.list()) {
-//            Log.info("@",fi);
-            }
-//        String serverLink = "https://w3g3a5v6.ssl.hwcdn.net/build/377240/1442306-archive-default?GoogleAccessId=uploader@moonscript2.iam.gserviceaccount.com&Expires=1624227195&Signature=bqM%2BK32GOMC7d1pA2KOd6xPjykHoyGHESSD5wraKzEnK%2BZf096h3lekhXg0615y0IkYIC%2BsqHVbDuOePan74h5TN42ZI0Jwo6UU%2B8%2FFVt%2BUUzVzfj8jhDZXPurs802PqNic7u1H9RdjnAl2ad2o98kBVYdOtuDc2v2KhhddnqT4dLxswIV1RdQjCtBtCnoJpt6wjWKL0Mo3RgiekHdcoswkZK1JlF%2B2fDfGzqRGHIBIVpU8yRpVNb8z3G%2BuO5u3jzVEHgoCnJr8glKMfacd9Wkoo%2Bn02EY0d0LZyO0lo3QuOBcA9AF%2FetLWfdtullTb2wpOO5ZbD1eb8KKN7J9DPWw==&hwexp=1624227455&hwsig=c472a60f0541e4cd5a1c2382fc52f0ce";
-//            String serverLink = "https://github.com/Anuken/Mindustry/releases/download/v126.2/server-release.jar";
-
-            for (Fi fi : compJava.list()) {
                 if (fi.isDirectory()) continue;
                 String className = fi.nameWithoutExtension();
                 selectedClassName = className;
-                if (className.equals("MinerComp")) {
-//                    Log.info("MinerComp\n@",fi.readString());
-                }
                 if (Seq.with("BuildingComp", "BulletComp", "DecalComp", "EffectStateComp", "FireComp", "LaunchCoreComp", "PlayerComp", "PuddleComp").contains(className)) {
-//                    Log.info("@ skipped", className);
                     compJava.child(fi.name()).delete();
                     finalComp.child(fi.name()).delete();
                     continue;
@@ -82,20 +76,18 @@ public class AnukeCompDownloader {
                     Log.info("@ skipped", className);
                     continue;
                 }
-
-//                Log.info("" + fi.name());
                 String file = fi.readString();
-                String convert = codeConverter.convert(file,className);
+                String convert = codeConverter.convert(file, className);
                 String string = convert
                         .replace("var core = team.core();", "mindustry.world.blocks.storage.CoreBlock.CoreBuild core = team.core();")
                         .replace("var core = core();", "mindustry.world.blocks.storage.CoreBlock.CoreBuild core = core();")
-                        .replace("var entry = statuses.find(e -> e.effect == effect);","StatusEntry entry = statuses.find(e -> e.effect == effect);")
-                        .replace("package mindustry.entities.comp;", "package gas.entities.compByAnuke;")
+                        .replace("var entry = statuses.find(e -> e.effect == effect);", "StatusEntry entry = statuses.find(e -> e.effect == effect);")
+                        .replace("package mindustry.entities.comp;", "package " + packageName + ".entities.compByAnuke;")
                         .replace("import static mindustry.logic.GlobalConstants.*;",
                                 "import static mindustry.logic.GlobalConstants.*;\n" + "import static mindustry.logic.LAccess.*;")
                         .replace("@Override", "__OVERRIDE__")
                         .replace("@Nullable", "__NULLABLE__")
-                        .replace("@", "@gas.annotations.GasAnnotations.")
+                        .replace("@", "@" + packageName + ".annotations." + annotationsClassName + ".")
                         .replace("__OVERRIDE__", "@Override")
                         .replace("__NULLABLE__", "@Nullable");
                 string = string.replace("};\n" +
@@ -103,35 +95,33 @@ public class AnukeCompDownloader {
                                                  "    }");
                 finalComp.child(fi.name()).writeString(string);
             }
-//            finalComp.copyTo(dir);
             Seq<String> names = new Seq<>();
             for (Fi fi : finalComp.list()) {
                 fi.copyTo(dir.child(fi.name()));
                 names.add(fi.nameWithoutExtension());
             }
             StringBuilder file = new StringBuilder();
-            file.append("package gas.entities.compByAnuke;\n" +
-                        "\n" +
-                        "import gas.annotations.GasAnnotations;\n" +
+            file.append("package " + packageName + ".entities.compByAnuke;\n\n" +
+                        "import " + packageName + ".annotations." + annotationsClassName + ";\n" +
                         "import mindustry.gen.Unitc;\n" +
                         "public class AnnotationConfigComponents {");
             for (String name : names) {
                 if (!name.endsWith("Comp")) continue;
                 String interfaceName = interfaceName(name);
-                file.append(Strings.format("@GasAnnotations.EntitySuperClass\n" +
+                file.append(Strings.format("@" + annotationsClassName + ".EntitySuperClass\n" +
                                            "    public static interface @ extends mindustry.gen.@{\n" +
                                            "    }", "@", interfaceName, interfaceName)).append("\n");
             }
             file.append("\n}");
             dir.child("AnnotationConfigComponents.java").writeString(file.toString());
             System.out.println(Strings.format("Time taken: @s", Time.nanosToMillis(Time.timeSinceNanos(nanos)) / 1000f));
-//        }
         } catch (Exception e) {
             e.printStackTrace();
         }
         folder.deleteDirectory();
-        folder.walk(f->f.delete());
+        folder.walk(f -> f.delete());
         folder.delete();
+        System.exit(0);
     }
 
     static String interfaceName(String comp) {
